Change to self-hosted only
Not possible to use the `ubuntu-latest` because of servers are located in the US. access restricted by the Binance platform.
`Service unavailable from a restricted location according to 'b. Eligibility' in https://www.binance.com/en/terms. Please contact customer service if you believe you received this message in error.`